test: Support both resource types

+++ test/test-out.sh
@@ -1,5 +1,11 @@
 #! /bin/bash
-request=$1
+type=$1
+request=$2
+
+if [[ -z $type || -z $request ]]; then
+    echo "Required arguments: <resource type> <request file>"
+    exit 1
+fi
 
 cat "$request" | docker run --rm -i \
 -e BUILD_NAME=mybuild \
@@ -7,4 +13,4 @@ cat "$request" | docker run --rm -i \
 -e BUILD_PIPELINE_NAME=mypipe \
 -e BUILD_TEAM_NAME=myteam \
 -e ATC_EXTERNAL_URL="https://example.com" \
--v "$(pwd)/out:/tmp/resource/out" jakobleben/slack-chat-resource /opt/resource/out /tmp/resource/out
+-v "$(pwd)/$type/out:/tmp/resource" jakobleben/slack-$type-resource /opt/resource/out /tmp/resource
